fix(test): resolve import cycle by moving GetStandardTestOptions to testopts package
Moved GetStandardTestOptions from internal/testing/util (which imports internal/plugin) to new internal/testing/testopts package to break circular dependency with test packages.
